[otbn,dv] Add a generator for badly nested loops
This generates two loops that share an end address, like the
following:

    loop   x2, 2
     loop  x2, 1
      nop

We'll execute the nop x2 times and then execution will continue, but
we won't pop the other loop off the stack.

This is an explicit functional coverage point (duplicate_loop_end_cp
in the insn_loop_cg and insn_loopi_cg coverage groups).

Getting this right in the generator was a little tricky because a
surrounding "Loop" generator no longer knows exactly what the state of
the loop stack is when its body finishes. I'm reasonably confident
that we track loop depths here properly because if it goes wrong, the
combination of the BadDeepLoop generator and end address tracking
soon flag up an error.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>