[spi_device] Refactor SRAM inteface
This commit bundles the SRAM signals in SPI_DEVICE HWIP into two data
structures, `sram_l1m_t` and `sram_m2l_t`. These bundles help the
SPI_DEVICE top to be well organized and clean.

In addition to the bundles, this commit defines the strobe for SRAM
write data. The strobe for FWMode and Read Command are tied to 1 as of
now. The SPI Flash Upload submodule is going to use the strobe to upload
the payload data in a byte granularity.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>