[otbn,dv] Add a generator for instructions using x1 more heavily
This is intended to generate things like "add x1, x1, x1" in more
situations. In particular, it can do so when the call stack is full.
That isn't possible for StraightLineInsn, because that picks operands
one at a time (this instruction is valid for a full call stack, but
"add x1, x0, x0" is not).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>