[otbn,dv] Fix callstack flags in coverage
pop_stack_a (resp. pop_stack_b) is signalled if the instruction in
question tries to read from x1 with operand a (resp b). pop_stack is
only true if, furthermore, rd_commit_i holds. This guarantees that the
instruction isn't stalled, but also implies that the instruction
didn't cause a fault.

We don't need to check for stalling instructions when calculating
these flags: they'll only be read in otbn_trace_monitor.sv if the
instruction isn't stalled. But we *do* need to make sure we get the
raw values for a faulting instruction. Without this fix, it would be
impossible to see e.g. how we caused an underflow with an instruction
like "add x0, x1, x1".

Similarly, the push_stack signal requires that the instruction didn't
cause a fault: we actually want the raw signal in push_stack_reqd.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>