[otbn,dv] Add loop warping support to ISS
This is the first commit for "loop warping". The idea is that you
might want to see the loop counter in the design incrementing at some
really big values (it goes up to 2^32), but don't want to simulate 4
billion cycles to get there.

Instead, you want to simulate the first few iterations around the loop
and then "jump" to the last few. This patch teaches the ISS to do this
when special symbols are added to the ELF file. Specifically, such a
symbol should have a name of the form:

    _loop_warp_123_456

and value that match an IMEM address. This means "if I execute the
given address and the innermost loop count is 123, alter the loop
count to 345".

We put the counts in the symbol name and address in the value so that
you can easily specify these in assembly code, without needing to know
the exact layout of the generated binary.

Of course, there's a duplication problem: what if you had two
different loops where you wanted to do the same warping? To get around
this, we allow any string after the second count:

    _loop_warp_123_456ANYOLDRUBBISH

and ignore it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>