[top, util] Add quick support for multiple power domain resets
This commits adds the ability for modules to be declared under
multiple power domains and thus assign resets of different domains
to each port.

Note this is a quick bolt on and not the ideal way of doing things.
Ideally, the concept of multiple domains should be expressed in the
ip.hjson themselves. The top level then hierarchically connects the
available top level domains to the sub-domains.

The ip.hjson would then also clearly state what local domain each
reset is related to, and the connective tissue should be built-up that
way instead of being mandated from the top.

However, that is a larger change than necessary at this point, and should
be accounted for later in the reset generation refactoring.

Signed-off-by: Timothy Chen <timothytim@google.com>