[fpga] Modify Boot ROM placement and checking
Previously, only the first two BRAM instances were checked, there
were only placement rules for the first four banks (with scrambling
there are five banks) and the placement seemed to be rather arbitrary
(the chosen sites were not available on some FPGAs).

With this commit, BRAM sites are chosen that are available on most FPGAs
and the placement/checking is also enabled on boards other than the
NexysVideo. Note this commit doesn't touch the splicing which currently
doesn't work with the scrambling enabled.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>