[checklist] Updates to checklist for D2 status
These changes are intended to reflect the outcome of a recent technical
committee discussion around the inclusion of security countermeasures in
D2 sign off.

Most blocks to go through D2 sign off so far have delayed final
implementation of security countermeasures to the D3 stage. This change
updates the checklist to reflect that reality, and gives some guidance
on which countermeasures would be ok to delay to D3 and which should be
included in D2.

Signed-off-by: Tom Roberts <tomroberts@lowrisc.org>