[rom_ctrl] Add utility script to generate splicable Vivado ROM images
Previously, a combination of srec_cat and a custom Python script was
was used to generate the Vivado ROM image from a .bin file. The new
script directly reads the .vmem file to produce the Vivado .mem file
usable for FPGA ROM splicing. But more importantly, this script is
also compatible with word widths different to 32 bits.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>