[otbn,doc] Add an "errs" field to instruction YAML
The idea is that this is used to document which errors the instruction
might cause when run. In theory, you can figure this out yourself from
the operation pseudo-code (mostly), but it seems like a good idea to
write it down explicitly.

This commit adds a list to ADD (the first instruction) and
ECALL (which happens to have none) as a demo.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>